# docs/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "BudgetOut": {
        "properties": {
          "category_id": {
            "title": "Category Id",
            "type": "string"
          },
          "left": {
            "title": "Left",
            "type": "integer"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "pct": {
            "title": "Pct",
            "type": "integer"
          },
          "spent": {
            "title": "Spent",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "category_id",
          "limit",
          "spent",
          "pct",
          "status",
          "left"
        ],
        "title": "BudgetOut",
        "type": "object"
      },
...
      "CategoryCreateIn": {
        "additionalProperties": false,
        "properties": {
          "color": {
            "title": "Color",
            "type": "string"
          },
          "icon": {
            "title": "Icon",
            "type": "string"
          },
          "monthly_limit": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 1000000000000.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monthly Limit"
          },
          "name": {
            "maxLength": 32,
            "title": "Name",
            "type": "string"
          },
          "type": {
            "$ref": "#/components/schemas/CategoryType"
          }
        },
        "required": [
          "name",
          "icon",
          "color",
          "type"
        ],
        "title": "CategoryCreateIn",
        "type": "object"
      },
      "CategoryDeleteIn": {
        "additionalProperties": false,
        "properties": {
          "reassign_to": {
            "anyOf": [
              {
                "maxLength": 36,
                "minLength": 2,
                "pattern": "^[a-z0-9\\-]+$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reassign To"
          }
        },
        "title": "CategoryDeleteIn",
        "type": "object"
      },
      "CategoryOut": {
        "properties": {
          "color": {
            "title": "Color",
            "type": "string"
          },
          "icon": {
            "title": "Icon",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_system": {
            "title": "Is System",
            "type": "boolean"
          },
          "monthly_limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monthly Limit"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "spent": {
            "title": "Spent",
            "type": "integer"
          },
          "transactions_count": {
            "title": "Transactions Count",
            "type": "integer"
          },
          "type": {
            "$ref": "#/components/schemas/CategoryType"
          }
        },
        "required": [
          "id",
          "name",
          "icon",
          "color",
          "type",
          "is_system",
          "monthly_limit",
          "spent",
          "transactions_count"
        ],
        "title": "CategoryOut",
        "type": "object"
      },
      "CategoryType": {
        "enum": [
          "expense",
          "income"
        ],
        "title": "CategoryType",
        "type": "string"
      },
      "CategoryUpdateIn": {
        "additionalProperties": false,
        "description": "Tizim kategoriyasida faqat monthly_limit va name (override) o'zgaradi.",
        "properties": {
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "monthly_limit": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 1000000000000.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monthly Limit"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 32,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "CategoryUpdateIn",
        "type": "object"
      },
...
      "ErrorOut": {
        "description": "Yagona xato formati (BE-1801). Qo'shimcha maydonlar koddan kelib chiqadi.",
        "properties": {
          "attempts_left": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attempts Left"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "fields": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "retry_after": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry After"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "title": "ErrorOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Finora API v1. Barcha vaqtlar ISO 8601 UTC; mijoz vaqt zonasi `X-Timezone` sarlavhasida. Summalar — so'm (butun son). Xato formati: `{code, message, request_id}` (+ `fields`, `retry_after`, `attempts_left`). Yozuv yaratuvchi POST'lar `Idempotency-Key` talab qiladi. Xato kodlari: docs/backend/ERROR_CODES.md.",
    "title": "Finora API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/budgets": {
      "get": {
        "description": "Limit qo'yilgan kategoriyalar: warning 0.75 ≤ pct ≤ 1, over pct > 1 (BE-1001).",
        "operationId": "list_budgets_v1_budgets_get",
        "parameters": [
          {
            "in": "query",
            "name": "month",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^\\d{4}-\\d{2}$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Month"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-timezone",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Timezone"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BudgetOut"
                  },
                  "title": "Response List Budgets V1 Budgets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "unauthorized / token_expired / session_expired"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "not_found (begona resurs ham)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "validation_error + fields[] va boshqa kodlar"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "rate_limited + retry_after"
          }
        },
        "summary": "List Budgets",
        "tags": [
          "budgets"
        ]
      }
    },
    "/v1/categories": {
      "get": {
        "description": "Tizim + user kategoriyalari, har birida joriy oy `spent` (BE-1501).",
        "operationId": "list_categories_v1_categories_get",
        "parameters": [
          {
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CategoryType"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "in": "header",
            "name": "x-timezone",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Timezone"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CategoryOut"
                  },
                  "title": "Response List Categories V1 Categories Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "unauthorized / token_expired / session_expired"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "not_found (begona resurs ham)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "validation_error + fields[] va boshqa kodlar"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "rate_limited + retry_after"
          }
        },
        "summary": "List Categories",
        "tags": [
          "categories"
        ]
      },
      "post": {
        "operationId": "create_category_v1_categories_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "unauthorized / token_expired / session_expired"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "not_found (begona resurs ham)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "validation_error + fields[] va boshqa kodlar"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "rate_limited + retry_after"
          }
        },
        "summary": "Create Category",
        "tags": [
          "categories"
        ]
      }
    },
    "/v1/categories/{category_id}": {
      "delete": {
        "description": "Faqat user kategoriyasi. Tranzaksiyalari bo'lsa `reassign_to` majburiy (409).",
        "operationId": "delete_category_v1_categories__category_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "category_id",
            "required": true,
            "schema": {
              "maxLength": 36,
              "minLength": 2,
              "pattern": "^[a-z0-9\\-]+$",
              "title": "Category Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/CategoryDeleteIn"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "unauthorized / token_expired / session_expired"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "not_found (begona resurs ham)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "validation_error + fields[] va boshqa kodlar"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "rate_limited + retry_after"
          }
        },
        "summary": "Delete Category",
        "tags": [
          "categories"
        ]
      },
      "patch": {
        "description": "`monthly_limit: null` — limitni olib tashlash (BE-1001).",
        "operationId": "update_category_v1_categories__category_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "category_id",
            "required": true,
            "schema": {
              "maxLength": 36,
              "minLength": 2,
              "pattern": "^[a-z0-9\\-]+$",
              "title": "Category Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CategoryUpdateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "unauthorized / token_expired / session_expired"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "not_found (begona resurs ham)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "validation_error + fields[] va boshqa kodlar"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "rate_limited + retry_after"
          }
        },
        "summary": "Update Category",
        "tags": [
          "categories"
        ]
      }
    },
...
  }
}